feat: add bot service layer
